requirements.txt
asgiref==3.4.1
Django==3.2.7
django-cors-headers==3.8.0
django-disqus==0.5
django-tagging==0.5.0
djangorestframework==3.12.4
mock==4.0.3
Pillow==8.3.2
pytz==2021.1
sqlparse==0.4.2
